.still wip event search

+++ module/Application/src/Application/memreas/AWSMemreasRedisCache.php
@@ -197,7 +197,7 @@ class AWSMemreasRedisCache {
 					/**
 					 * - Check for viewable from / to since db date is string
 					 */
-					if ((! empty ( trim ( $event ['viewable_from'] ) )) && (! empty ( trim ( $event ['viewable_to'] ) ))) {
+					if ((! empty ( ( int ) (trim ( $event ['viewable_from'] )) )) && (! empty ( ( int ) (trim ( $event ['viewable_to'] )) ))) {
 						$now = time ();
 						/**
 						 * check if now is outside dates
@@ -210,12 +210,12 @@ class AWSMemreasRedisCache {
 					/**
 					 * - Check for self_destruct/ghost since db date is string
 					 */
-					if (! empty ( trim ( $event ['self_destuct'] ) )) {
+					if (! empty ( ( int ) (trim ( $event ['self_destuct'] )) )) {
 						$now = time ();
 						/**
 						 * check if now is outside dates
 						 */
-						Mlog::addone('Check for ghost date--->', '$now--->'.$now.' $ghost');
+						Mlog::addone ( 'Check for ghost date--->', '$now--->' . $now . ' $ghost' );
 						if ($now <= $event ['self_destruct']) {
 							Mlog::addone ( "warmMemreasSet::checking ghost is out of bounds for now $now --> ", $event ['event_id'] );
 							continue;
